Move legacy package inside flows/definition
