+++ flows/definition/legacy/compat.go
