+++ flows/definition/legacy/compat_test.go
@@ -3,7 +3,7 @@ package legacy_test
 import (
 	"testing"
 
-	"github.com/nyaruka/goflow/legacy"
+	"github.com/nyaruka/goflow/flows/definition/legacy"
 	"github.com/nyaruka/goflow/test"
 
 	"github.com/stretchr/testify/assert"
