+++ flows/definition/legacy/expressions/migrate.go
@@ -10,7 +10,7 @@ import (
 	"github.com/nyaruka/goflow/envs"
 	"github.com/nyaruka/goflow/excellent"
 	"github.com/nyaruka/goflow/flows"
-	"github.com/nyaruka/goflow/legacy/gen"
+	"github.com/nyaruka/goflow/flows/definition/legacy/gen"
 
 	"github.com/antlr/antlr4/runtime/Go/antlr"
 )
